# Conflicts:
#	lib/Varien/Data/Form/Element/Checkboxes.php

+++ app/code/core/Mage/Admin/Model/Resource/User.php
@@ -14,6 +14,9 @@
  */
 class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstract
 {
+    /**
+     * @inheritDoc
+     */
     protected function _construct()
     {
         $this->_init('admin/user', 'user_id');
@@ -41,6 +44,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
      * Authenticate user by $username and $password
      *
      * @return $this
+     * @throws Zend_Db_Adapter_Exception
      */
     public function recordLogin(Mage_Admin_Model_User $user)
     {
@@ -63,8 +67,8 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Load data by specified username
      *
-     * @param string $username
-     * @return false|array
+     * @param  string      $username
+     * @return array|false
      */
     public function loadByUsername($username)
     {
@@ -84,7 +88,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Check if user is assigned to any role
      *
-     * @param int|Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
+     * @param  int|Mage_Admin_Model_User|Mage_Core_Model_Abstract|string $user
      * @return null|array
      */
     public function hasAssigned2Role($user)
@@ -119,53 +123,57 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Set created/modified values before user save
      *
-     * @param Mage_Admin_Model_User $user
+     * @param Mage_Admin_Model_User $object
      * @inheritDoc
      */
-    protected function _beforeSave(Mage_Core_Model_Abstract $user)
+    protected function _beforeSave(Mage_Core_Model_Abstract $object)
     {
-        if ($user->isObjectNew()) {
-            $user->setCreated($this->formatDate(true));
+        if ($object->isObjectNew()) {
+            $object->setCreated($this->formatDate(true));
         }
 
-        $user->setModified($this->formatDate(true));
+        $object->setModified($this->formatDate(true));
 
-        return parent::_beforeSave($user);
+        return parent::_beforeSave($object);
     }
 
     /**
      * Unserialize user extra data after user save
      *
+     * @param  Mage_Admin_Model_User $object
      * @return $this
      */
-    protected function _afterSave(Mage_Core_Model_Abstract $user)
+    protected function _afterSave(Mage_Core_Model_Abstract $object)
     {
-        $this->_unserializeExtraData($user);
+        $this->_unserializeExtraData($object);
         return $this;
     }
 
     /**
      * Unserialize user extra data after user load
      *
+     * @param Mage_Admin_Model_User $object
      * @inheritDoc
      */
-    protected function _afterLoad(Mage_Core_Model_Abstract $user)
+    protected function _afterLoad(Mage_Core_Model_Abstract $object)
     {
-        return parent::_afterLoad($this->_unserializeExtraData($user));
+        return parent::_afterLoad($this->_unserializeExtraData($object));
     }
 
     /**
      * Delete user role record with user
      *
+     * @param  Mage_Admin_Model_User $object
      * @return $this
      * @throws Exception
+     * @throws Throwable
      */
-    public function delete(Mage_Core_Model_Abstract $user)
+    public function delete(Mage_Core_Model_Abstract $object)
     {
-        $this->_beforeDelete($user);
+        $this->_beforeDelete($object);
         $adapter = $this->_getWriteAdapter();
 
-        $uid = $user->getId();
+        $uid = $object->getId();
         $adapter->beginTransaction();
         try {
             $conditions = [
@@ -180,14 +188,17 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             throw $throwable;
         }
 
-        $this->_afterDelete($user);
+        $this->_afterDelete($object);
         return $this;
     }
 
     /**
      * TODO: unify _saveRelations() and add() methods, they make same things
      *
-     * @return $this|Mage_Core_Model_Abstract
+     * @param  Mage_Admin_Model_User                                $user
+     * @return $this|Mage_Admin_Model_User|Mage_Core_Model_Abstract
+     * @throws Mage_Core_Exception
+     * @throws Zend_Db_Adapter_Exception
      */
     public function _saveRelations(Mage_Core_Model_Abstract $user)
     {
@@ -232,9 +243,12 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             }
 
             $adapter->commit();
-        } catch (Mage_Core_Exception|Exception $e) {
+        } catch (Mage_Core_Exception $mageCoreException) {
             $adapter->rollBack();
-            throw $e;
+            throw $mageCoreException;
+        } catch (Exception $exception) {
+            $adapter->rollBack();
+            throw $exception;
         }
 
         return $this;
@@ -243,6 +257,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Get user roles
      *
+     * @param  Mage_Admin_Model_User $user
      * @return array
      */
     public function getRoles(Mage_Core_Model_Abstract $user)
@@ -257,10 +272,10 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                     ->from($table, [])
                     ->joinLeft(
                         ['ar' => $table],
-                        "(ar.role_id = {$table}.parent_id and ar.role_type = 'G')",
+                        "(ar.role_id = $table.parent_id and ar.role_type = 'G')",
                         ['role_id'],
                     )
-                    ->where("{$table}.user_id = :user_id");
+                    ->where("$table.user_id = :user_id");
 
         $binds = [
             'user_id' => (int) $user->getId(),
@@ -278,13 +293,16 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Save user roles
      *
+     * @param  Mage_Admin_Model_User     $user
      * @return $this
+     * @throws Zend_Cache_Exception
+     * @throws Zend_Db_Adapter_Exception
      */
     public function add(Mage_Core_Model_Abstract $user)
     {
         $dbh = $this->_getWriteAdapter();
         $aRoles = $this->hasAssigned2Role($user);
-        if (count($aRoles)) {
+        if ($aRoles && count($aRoles)) {
             foreach ($aRoles as $data) {
                 $dbh->delete(
                     $this->getTable('admin/role'),
@@ -322,6 +340,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Delete user role
      *
+     * @param  Mage_Admin_Model_User $user
      * @return $this
      */
     public function deleteFromRole(Mage_Core_Model_Abstract $user)
@@ -348,6 +367,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Check if role user exists
      *
+     * @param  Mage_Admin_Model_User $user
      * @return array
      */
     public function roleUserExists(Mage_Core_Model_Abstract $user)
@@ -375,6 +395,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Check if user exists
      *
+     * @param  Mage_Admin_Model_User $user
      * @return array|false
      */
     public function userExists(Mage_Core_Model_Abstract $user)
@@ -398,9 +419,10 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Save user extra data
      *
-     * @param Mage_Core_Model_Abstract $object
-     * @param string $data
+     * @param  Mage_Admin_Model_User     $object
+     * @param  string                    $data
      * @return $this
+     * @throws Zend_Db_Adapter_Exception
      */
     public function saveExtra($object, $data)
     {
@@ -418,9 +440,11 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Set reload ACL flag
      *
-     * @param Mage_Core_Model_Abstract $object
-     * @param int $flag
+     * @param  Mage_Admin_Model_User     $object
+     * @param  int                       $flag
      * @return $this
+     * @throws Zend_Cache_Exception
+     * @throws Zend_Db_Adapter_Exception
      */
     public function saveReloadAclFlag($object, $flag)
     {
@@ -445,6 +469,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Unserializes user extra data
      *
+     * @param  Mage_Admin_Model_User    $user
      * @return Mage_Core_Model_Abstract
      */
     protected function _unserializeExtraData(Mage_Core_Model_Abstract $user)
